+++ core/server/services/route-settings/validate.js
@@ -429,7 +429,8 @@ module.exports = function validate(object) {
 
     debug('api version', apiVersion);
 
-    RESOURCE_CONFIG = require(`../routing/config/${apiVersion}`);
+    // TODO: extract this config outta here! the config should be passed into this module
+    RESOURCE_CONFIG = require(`../../../frontend/services/routing/config/${apiVersion}`);
 
     object.routes = _private.validateRoutes(object.routes);
     object.collections = _private.validateCollections(object.collections);
