+++ test/unit/services/settings/yaml-parser.test.js
@@ -3,7 +3,7 @@ const should = require('should');
 const fs = require('fs-extra');
 const yaml = require('js-yaml');
 const path = require('path');
-const yamlParser = require('../../../../core/frontend/services/settings/yaml-parser');
+const yamlParser = require('../../../../core/server/services/route-settings/yaml-parser');
 
 describe('UNIT > Settings Service yaml parser:', function () {
     let yamlSpy;
